Refactored code into separate MapFragment class. Added a floating action button for creating a new report.
